# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/system/audit": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 500,
              "default": 200
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "description": "逗号分隔的多值，取值 info / warn / error，服务端按 IN 过滤；单值向后兼容（`severity=error`），「仅告警」= `severity=warn,error`；重复值去重，含非法值 → 400 VALIDATION_FAILED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "sandbox",
                "project",
                "credential",
                "image",
                "system"
              ],
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO 时间上界（与游标正交）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO 时间下界（与游标正交）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "description": "向老翻页：返回 seq < before",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "since",
            "required": false,
            "in": "query",
            "description": "向新翻页：返回 seq > since",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponseDto"
                }
              }
            }
          }
        },
        "summary": "平台级审计流。since/before 为按 seq 的双向游标（互斥，同传 400 VALIDATION_FAILED）；from/to 是与游标正交的时间过滤。响应恒按 seq 降序 + hasMore。",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/automations/{id}/deletion-preview": {
      "get": {
        "operationId": "AutomationController_deletionPreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationDeletionPreviewResponseDto"
                }
              }
            }
          }
        },
        "summary": "Read authoritative run, artifact and active task deletion consequences",
        "tags": [
          "automation"
        ]
      }
    },
...
    "/api/automations/webhook-test": {
      "post": {
        "operationId": "AutomationController_webhookTest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookTestRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookTestResultDto"
                }
              }
            }
          }
        },
        "summary": "Send a sample webhook payload (same timeout + SSRF policy)",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/automations/runs/{runId}": {
      "get": {
        "operationId": "AutomationController_getRun",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationRunResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get one automation run",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/automations/runs/{runId}/logs": {
      "get": {
        "operationId": "AutomationController_readRunLogs",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "absent ⇒ the trailing `limit` bytes",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "default 65536, max 1048576",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw log bytes as UTF-8 text",
            "headers": {
              "x-log-offset": {
                "description": "first byte of this slice",
                "schema": {
                  "type": "integer"
                }
              },
              "x-log-total": {
                "description": "total bytes on disk",
                "schema": {
                  "type": "integer"
                }
              },
              "x-log-eof": {
                "description": "\"true\" when this slice reaches EOF",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Read a byte range of a run log (defaults to the trailing 64KB)",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/automations/{id}": {
      "get": {
        "operationId": "AutomationController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get an automation rule",
        "tags": [
          "automation"
        ]
      },
      "put": {
        "operationId": "AutomationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAutomationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Update a rule — omitting `timezone` keeps the snapshot (I-AUT-9)",
        "tags": [
          "automation"
        ]
      },
      "delete": {
        "operationId": "AutomationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Delete a rule (its run history cascades)",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/automations/{id}/enable": {
      "post": {
        "operationId": "AutomationController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Re-enable a rule (clears failure count and the degraded flag)",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/automations/{id}/disable": {
      "post": {
        "operationId": "AutomationController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Disable a rule",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/automations/{id}/runs": {
      "get": {
        "operationId": "AutomationController_listRuns",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "description": "run id; returns strictly older ones",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "default 20, max 100",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAutomationRunsDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "List a rule’s run history, newest first (cursor)",
        "tags": [
          "automation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agent Platform API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "AuditListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "seq": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true
                },
                "at": {
                  "type": "string",
                  "format": "date-time"
                },
                "category": {
                  "type": "string",
                  "enum": [
                    "sandbox",
                    "project",
                    "credential",
                    "image",
                    "system"
                  ]
                },
                "type": {
                  "type": "string"
                },
                "severity": {
                  "type": "string",
                  "enum": [
                    "info",
                    "warn",
                    "error"
                  ]
                },
                "subjectType": {
                  "type": "string"
                },
                "subjectId": {
                  "type": "string"
                },
                "actor": {
                  "type": "string",
                  "minLength": 1
                },
                "summary": {
                  "type": "string"
                },
                "detail": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "durationMs": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false
                },
                "outcome": {
                  "type": "string",
                  "enum": [
                    "ok",
                    "failed",
                    "skipped"
                  ]
                },
                "errorCode": {
                  "type": "string"
                }
              },
              "required": [
                "seq",
                "at",
                "category",
                "type",
                "severity",
                "actor",
                "summary"
              ]
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
...
      "AutomationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "runtime": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "scheduleKind": {
            "type": "string",
            "enum": [
              "hourly",
              "daily",
              "weekly"
            ]
          },
          "scheduleConfig": {
            "type": "object",
            "properties": {
              "minute": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": false,
                "maximum": 59,
                "exclusiveMaximum": false
              },
              "time": {
                "type": "string",
                "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
              },
              "days": {
                "type": "array",
                "minItems": 1,
                "maxItems": 7,
                "items": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false,
                  "maximum": 6,
                  "exclusiveMaximum": false
                }
              }
            }
          },
          "timezone": {
            "type": "string"
          },
          "timeoutMinutes": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 30,
                "maximum": 30
              },
              {
                "type": "number",
                "minimum": 60,
                "maximum": 60
              },
              {
                "type": "number",
                "minimum": 120,
                "maximum": 120
              },
              {
                "type": "number",
                "minimum": 240,
                "maximum": 240
              }
            ]
          },
          "artifactRetentionDays": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 3,
                "maximum": 3
              },
              {
                "type": "number",
                "minimum": 7,
                "maximum": 7
              },
              {
                "type": "number",
                "minimum": 30,
                "maximum": 30
              }
            ]
          },
          "webhookUrl": {
            "type": "string",
            "format": "uri"
          },
          "triggerOn": {
            "type": "string",
            "enum": [
              "failure",
              "success",
              "all"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "degraded": {
            "type": "boolean"
          },
          "consecutiveFailures": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "lastTriggeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "nextTriggerAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "runtime",
          "prompt",
          "scheduleKind",
          "scheduleConfig",
          "timezone",
          "timeoutMinutes",
          "artifactRetentionDays",
          "triggerOn",
          "enabled",
          "degraded",
          "consecutiveFailures",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AutomationDeletionPreviewResponseDto": {
        "type": "object",
        "properties": {
          "runCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "artifactCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "runningTasks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name"
              ]
            }
          }
        },
        "required": [
          "runCount",
          "artifactCount",
          "runningTasks"
        ]
      },
...
      "WebhookTestRequestDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          }
        },
        "required": [
          "url"
        ]
      },
      "WebhookTestResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "HOST_NOT_ALLOWED",
              "TIMEOUT",
              "UPSTREAM_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      },
      "AutomationRunResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "automationId": {
            "type": "string"
          },
          "sandboxId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "success",
              "failed",
              "timeout",
              "resource-exhausted",
              "skipped",
              "missed"
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "PREVIOUS_RUNNING",
              "AUTH_EXPIRED",
              "RESOURCE_EXHAUSTED"
            ]
          },
          "errorMessage": {
            "type": "string"
          },
          "retryCount": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 5,
            "exclusiveMaximum": false
          },
          "retryAt": {
            "type": "string",
            "format": "date-time"
          },
          "triggeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMs": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "outputSummary": {
            "type": "string"
          },
          "webhookStatus": {
            "type": "string",
            "enum": [
              "sent",
              "failed",
              "skipped"
            ]
          }
        },
        "required": [
          "id",
          "automationId",
          "status",
          "retryCount",
          "triggeredAt"
        ]
      },
      "UpdateAutomationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "runtime": {
            "type": "string",
            "minLength": 1
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          },
          "scheduleKind": {
            "type": "string",
            "enum": [
              "hourly",
              "daily",
              "weekly"
            ]
          },
          "scheduleConfig": {
            "type": "object",
            "properties": {
              "minute": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": false,
                "maximum": 59,
                "exclusiveMaximum": false
              },
              "time": {
                "type": "string",
                "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
              },
              "days": {
                "type": "array",
                "minItems": 1,
                "maxItems": 7,
                "items": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false,
                  "maximum": 6,
                  "exclusiveMaximum": false
                }
              }
            }
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "timeoutMinutes": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 30,
                "maximum": 30
              },
              {
                "type": "number",
                "minimum": 60,
                "maximum": 60
              },
              {
                "type": "number",
                "minimum": 120,
                "maximum": 120
              },
              {
                "type": "number",
                "minimum": 240,
                "maximum": 240
              }
            ]
          },
          "artifactRetentionDays": {
            "oneOf": [
              {
                "type": "number",
                "minimum": 3,
                "maximum": 3
              },
              {
                "type": "number",
                "minimum": 7,
                "maximum": 7
              },
              {
                "type": "number",
                "minimum": 30,
                "maximum": 30
              }
            ]
          },
          "webhookUrl": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "triggerOn": {
            "type": "string",
            "enum": [
              "failure",
              "success",
              "all"
            ]
          }
        }
      },
      "PaginatedAutomationRunsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "automationId": {
                  "type": "string"
                },
                "sandboxId": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "running",
                    "success",
                    "failed",
                    "timeout",
                    "resource-exhausted",
                    "skipped",
                    "missed"
                  ]
                },
                "errorCode": {
                  "type": "string",
                  "enum": [
                    "PREVIOUS_RUNNING",
                    "AUTH_EXPIRED",
                    "RESOURCE_EXHAUSTED"
                  ]
                },
                "errorMessage": {
                  "type": "string"
                },
                "retryCount": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false,
                  "maximum": 5,
                  "exclusiveMaximum": false
                },
                "retryAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "triggeredAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "startedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "completedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "durationMs": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false
                },
                "outputSummary": {
                  "type": "string"
                },
                "webhookStatus": {
                  "type": "string",
                  "enum": [
                    "sent",
                    "failed",
                    "skipped"
                  ]
                }
              },
              "required": [
                "id",
                "automationId",
                "status",
                "retryCount",
                "triggeredAt"
              ]
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
...
    }
  }
}